helpz to nourish hope

+++ client/src/components/navbar/NavBar.jsx
@@ -1,34 +1,46 @@
-import React, { useState,useEffect } from 'react'
+import React, { useState, useEffect } from 'react'
 import axios from 'axios'
+import Swal from 'sweetalert2'
 export default function NavBar() {
-    
+
     const [role, setRole] = useState(localStorage.getItem('role'))
     console.log(role);
     //
     const logOut = () => {
-        localStorage.clear()
-        window.location.href = '/'
+        Swal.fire({
+            title: "Do you want logout?",
+            showDenyButton: true,
+            confirmButtonText: "Yes",
+            denyButtonText: `No`
+        }).then((result) => {
+            /* Read more about isConfirmed, isDenied below */
+            if (result.isConfirmed) {
+                localStorage.clear()
+                window.location.href = '/'
+            } 
+        });
+
     }
     const [listDonations, setListDonations] = useState([])
     console.log(listDonations);
-    
-    console.log(listDonations.filter((item) =>item?.orphanage_id?.some((data) => data?.status === 0)).length );
 
-  
-  
+    console.log(listDonations.filter((item) => item?.orphanage_id?.some((data) => data?.status === 0)).length);
+
+
+
     useEffect(() => {
-      const login_id = localStorage.getItem('loginId')
-      axios.get(`http://localhost:5000/api/rest/list_donations_user/${login_id}`).then((res) => {
-        setListDonations(res.data.data)
-      })
+        const login_id = localStorage.getItem('loginId')
+        axios.get(`http://localhost:5000/api/rest/list_donations_user/${login_id}`).then((res) => {
+            setListDonations(res.data.data)
+        })
     }, [])
     const changeStatus = () => {
         console.log('looo');
-        const login_id = localStorage.getItem('loginId')    
+        const login_id = localStorage.getItem('loginId')
         axios.put(`http://localhost:5000/api/rest/update_donations/${login_id}`).then((res) => {
             console.log('updated');
-            
-          })    
+
+        })
         window.location.href = '/view-donation-history'
     }
 
@@ -39,7 +51,7 @@ export default function NavBar() {
                 <div className="navbar navbar-expand-lg bg-dark navbar-dark">
                     <div className="container-fluid">
                         <a href="index.html" className="navbar-brand">
-                            Helpz
+                            Nourish Hope
                         </a>
                         <button
                             type="button"
@@ -60,23 +72,23 @@ export default function NavBar() {
                                 <a href="/restaurants" className="nav-item nav-link">
                                     Restaurants
                                 </a>
-                                <a  onClick={changeStatus} style={{position:'relative'}} className="nav-item nav-link">
+                                <a onClick={changeStatus} style={{ position: 'relative' }} className="nav-item nav-link">
                                     Donations
-                                    {listDonations.filter((item) =>item?.orphanage_id?.some((data) => data?.status === 0)).length>0?
-                                    <span style={{
-                                        position:'absolute',
-                                        top:'0',
-                                        right:'0',
-                                        backgroundColor:'red', 
-                                        borderRadius:'30%',
-                                        width:'20px',
-                                        display:'flex',
-                                        justifyContent:'center',
-                                        alignItems:'end',
-                                        padding:'2px',
-                                        }}>{listDonations.filter((item) =>item?.orphanage_id?.some((data) => data?.status === 0)).length}</span>: null
+                                    {listDonations.filter((item) => item?.orphanage_id?.some((data) => data?.status === 0)).length > 0 ?
+                                        <span style={{
+                                            position: 'absolute',
+                                            top: '0',
+                                            right: '0',
+                                            backgroundColor: 'red',
+                                            borderRadius: '30%',
+                                            width: '20px',
+                                            display: 'flex',
+                                            justifyContent: 'center',
+                                            alignItems: 'end',
+                                            padding: '2px',
+                                        }}>{listDonations.filter((item) => item?.orphanage_id?.some((data) => data?.status === 0)).length}</span> : null
                                     }
-                                    
+
                                 </a>
                                 {/* <a href="/notification" style={{position:'relative'}} className="nav-item nav-link">
                                     Notification <span style={{
